chore: clean repo
