--- devutils/benchs/kokkos_omp_cmake_init
@@ -1,7 +1,0 @@
-#!/bin/bash 
-
-# TO USE IN KOKKOS ROOT 
-
-mkdir -p build 
-
-cmake -B build -S . -DKokkos_ENABLE_OPENMP=OFF -DKokkos_ENABLE_THREADS=ON -DCMAKE_BUILD_TYPE=Release -DKokkos_ENABLE_HWLOC=ON -DCMAKE_CXX_COMPILER=^Cang++-18 -DKokkos_ENABLE_CXX20=ON -DCMAKE_CXX_STANDARD=20 -DCMAKE_POSITION_INDEPENDENT_CODE=ON -DCMAKE_INSTALL_PREFIX=../kokkos
