--- devutils/datamodel/input_scheme.xml
@@ -1,23 +1,0 @@
-   <?xml version="1.0" encoding="UTF-8"?>
-   <cases>
-      <control name="n14">
-         <cma_case_path recursive="true">/path/to/cfd/flowmap_14/</cma_case_path>
-         <final_time>28000</final_time>
-         <number_particle>1000</number_particle>
-         <delta_time>0.01</delta_time>
-         <results_file_name>monod_14_long_3</results_file_name>
-         <number_exported_result>100</number_exported_result>
-         <model_name>model_monod</model_name>
-      </control>
-      
-      <control name="b20l_lb">
-         <cma_case_path >./cma_data/b20l/</cma_case_path>
-         <final_time>36</final_time>
-         <number_particle>1000</number_particle>
-         <delta_time>0</delta_time>
-         <results_file_name>lb_b20l_nofeed</results_file_name>
-         <number_exported_result>50</number_exported_result>
-         <model_name>model_lb</model_name>
-      </control>
-
-   </cases>
