--- devutils/datamodel/result_scheme.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0"?>
-
-<attributes>
-  <author>Autho</author>
-  <creationDate>today</creationDate>
-  <description>foo</description>
-</attributes>
-
-<misc>
-  <n_nodes>1</n_nodes>
-  <n_thread>8</n_thread>
-</misc>
-
-<final_result>
-  <number_particles>1001</number_particles>
-  <concentrations>
-    <liquid>
-      matrice 3d (ntimestep,nspecies,ncompartment)
-    </liquid>
-    <gas>
-      matrice 3d (ntimestep,nspecies,ncompartment)
-    </gas>
-    <particle_number>
-      matrice 3d (ntimestep,nspecies,ncompartment)
-    </particle_number>
-  </concentrations>
-</final_result>
-
-<initial_parameter>
-  <delta_time>0.1</delta_time>
-  <number_particles>1000</number_particles>
-  <n_map>14</n_map>
-  <particle_distribution>
-    array of int
-  </particle_distribution>
-  <final_time>10</final_time>
-  <number_compartment>500</number_compartment>
-</initial_parameter>
