--- devutils/datamodel/result_scheme.xsd
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema">
-
-  <xs:element name="attributes">
-    <xs:complexType>
-      <xs:sequence>
-        <xs:element name="author" type="xs:string"/>
-        <xs:element name="creationDate" type="xs:string"/>
-        <xs:element name="description" type="xs:string"/>
-      </xs:sequence>
-    </xs:complexType>
-  </xs:element>
-
-  <xs:element name="misc">
-    <xs:complexType>
-      <xs:sequence>
-        <xs:element name="n_nodes" type="xs:integer"/>
-        <xs:element name="n_thread" type="xs:integer"/>
-      </xs:sequence>
-    </xs:complexType>
-  </xs:element>
-
-  <xs:element name="final_result">
-    <xs:complexType>
-      <xs:sequence>
-        <xs:element name="number_particles" type="xs:integer"/>
-        <xs:element name="concentrations">
-          <xs:complexType>
-            <xs:sequence>
-              <xs:element name="liquid" type="xs:double"/>
-              <xs:element name="gas" type="xs:double"/>
-              <xs:element name="particle_number" type="xs:integer"/>
-            </xs:sequence>
-          </xs:complexType>
-        </xs:element>
-      </xs:sequence>
-    </xs:complexType>
-  </xs:element>
-
-  <xs:element name="initial_parameter">
-    <xs:complexType>
-      <xs:sequence>
-        <xs:element name="delta_time" type="xs:decimal"/>
-        <xs:element name="number_particles" type="xs:integer"/>
-        <xs:element name="n_map" type="xs:integer"/>
-        <xs:element name="final_time" type="xs:decimal"/>
-        <xs:element name="number_compartment" type="xs:integer"/>
-      </xs:sequence>
-    </xs:complexType>
-  </xs:element>
-
-</xs:schema>
