--- devutils/datamodel/toml/phase_descriptor.toml
@@ -1,32 +1,0 @@
-[Species]
-Liquid = ["glucose", "oxygen"]
-Gas = ["oxygen"]
-
-[Initial condition]
-Liquid = ["./path_to_h5_liq"]
-Gas = ["./path_to_h5_gas"]
-
-[[Feed]]
-flow_value = 1.0             
-concentration = [0.6]               
-position = [0,1]               
-species = [0]                
-props = "Constant"           
-
-[[Feed]]
-flow_value = 1.0             
-concentration = [0.6]               
-position = [0,1]               
-species = [0]                
-props = "Pulse"   
-start_time = 10
-end_time = 20
-pulse: 1        
-
-
-[[Feed]]
-flow_value = 0.03813511651379644  # Flow value for the gas phase
-concentration = [0.21]               # Vector of feed target values (e.g., oxygen concentration)
-position = [0]                # Vector of positions where the feed is applied
-species = [1]                 # Vector of species indices (e.g., 1 for oxygen in gas phase)
-props = "Constant"            # Feed type properties (Constant feed)
