--- docker-compose.yml
@@ -1,67 +1,0 @@
-# version: "3.6"
-services:
-  biocma_app:
-    &biocma_app
-    # image: bcasale/bio_cma_mc_runtime
-    build: 
-      context: ./devutils/docker
-      dockerfile: Dockerfile
-    environment: 
-      - CXX=clang++-18
-    container_name: biocma_mcst_dev
-    volumes:
-      - ./devutils/docker/runtime_container:/app:Z
-      - ./apps/:/app/apps:Z
-      # - ./subprojects:/app/subprojects:Z
-      # - ./modules:/app/modules:Z
-      - ./subprojects/dynlib.wrap:/app/subprojects/dynlib.wrap
-      - ./subprojects/cereal.wrap:/app/subprojects/cereal.wrap
-      - ./subprojects/highfive.wrap:/app/subprojects/highfive.wrap
-      - ./subprojects/cmtool.wrap:/app/subprojects/cmtool.wrap
-      - ./subprojects/cmtool:/app/subprojects/cmtool:Z
-      - ./cma_data/:/app/cma_data:Z
-      - ./meson.build:/app/meson.build:Z
-      - ./meson_options.txt:/app/meson_options.txt:Z
-      - ./devutils/auto_generate:/app/devutils/auto_generate:Z
-      - ./devutils/kokkos_wrap:/app/devutils/kokkos_wrap:Z
-      - ./devutils/datamodel:/app/devutils/datamodel:Z
-      - ./devutils/exec.py:/app/devutils/exec.py:Z
-      - ./tools/cli_formater.py:/app/tools/cli_formater.py:Z
-      - ./tools/runner.py:/app/tools/runner.py:Z
-      - ./tools/cases.xml:/app/tools/cases.xml:Z
-
-  biocma_app_nvidia_test:
-    <<: *biocma_app
-    runtime: nvidia
-    environment:
-      - NVIDIA_VISIBLE_DEVICES=all
-      - NVIDIA_DRIVER_CAPABILITIES=compute,utility
-  
-  biocma_app_cuda:
-    <<: *biocma_app
-    runtime: nvidia
-    environment:
-      - NVIDIA_VISIBLE_DEVICES=all
-      - NVIDIA_DRIVER_CAPABILITIES=compute,utility
-      - CXX=/app/devutils/kokkos_wrap/wrap_cxx
-    command:
-    - /bin/sh
-    - -c
-    - | 
-      rm -rf ./builddir 
-      meson setup builddir -Duse_cuda=true -Duse_system_kokkos=true -Dbuildtype=release
-       ninja -C builddir 
-
-  biocma_app_task_test:
-    <<: *biocma_app
-    container_name: biocma_mcst_dev_test
-    environment: 
-      - CXX=clang++-19
-    command:
-      - /bin/sh
-      - -c
-      - |
-        rm -rf ./builddir 
-        meson setup builddir --buildtype=debug -Dbuild_test=true -Duse_system_kokkos=true -ci_execution=true --reconfigure
-        ninja -C builddir  test 
-
