+++ documentation/.gitignore
@@ -2,3 +2,4 @@ html/
 latex/ 
 build/
 output/
+
