--- documentation/cli/main-blx.bib
@@ -1,11 +1,0 @@
-@Comment{$ biblatex control file $}
-@Comment{$ biblatex bbl format version 3.0 $}
-% Do not modify this file!
-%
-% This is an auxiliary file used by the 'biblatex' package.
-% This file may safely be deleted. It will be recreated as
-% required.
-
-@Control{biblatex-control,
-  options = {3.0:0:0:1:0:1:1:0:0:1:0:2:2:1:3:1:79:+:+:nyt},
-}
