--- documentation/cli/main.bbl
@@ -1,20 +1,0 @@
-% $ biblatex auxiliary file $
-% $ biblatex bbl format version 3.0 $
-% Do not modify the above lines!
-%
-% This is an auxiliary file used by the 'biblatex' package.
-% This file may safely be deleted. It will be recreated as
-% required.
-%
-\begingroup
-\makeatletter
-\@ifundefined{ver@biblatex.sty}
-  {\@latex@error
-     {Missing 'biblatex' package}
-     {The bibliography requires the 'biblatex' package.}
-      \aftergroup\endinput}
-  {}
-\endgroup
-
-\datalist[entry]{nyt/global//global/global}\enddatalist
-\endinput
