--- documentation/cli/main.blg
@@ -1,10 +1,0 @@
-This is BibTeX, Version 0.99dThe top-level auxiliary file: main.aux
-The style file: biblatex.bst
-I couldn't open database file ../common_ref.bib
----line 20 of file main.aux
- : \bibdata{main-blx,../common_ref
- :                                ,references}
-I'm skipping whatever remains of this command
-Database file #1: main-blx.bib
-Biblatex version: 3.11
-(There was 1 error message)
