--- documentation/cli/main.toc
@@ -1,9 +1,0 @@
-\boolfalse {citerequest}\boolfalse {citetracker}\boolfalse {pagetracker}\boolfalse {backtracker}\relax 
-\defcounter {refsection}{0}\relax 
-\contentsline {subsection}{\numberline {0.1}Basic Command Structure}{1}{subsection.0.1}% 
-\defcounter {refsection}{0}\relax 
-\contentsline {section}{List of Figures}{2}{subsection.0.1}% 
-\defcounter {refsection}{0}\relax 
-\contentsline {section}{List of Tables}{2}{section*.3}% 
-\defcounter {refsection}{0}\relax 
-\contentsline {section}{References}{2}{section*.4}%
