--- documentation/models/texput.log
@@ -1,19 +1,0 @@
-This is pdfTeX, Version 3.14159265-2.6-1.40.20 (MiKTeX 2.9.6960) (preloaded format=pdflatex 2023.12.22)  3 SEP 2024 16:55
-entering extended mode
-**main.tex
-
-! Emergency stop.
-<*> main.tex
-            
-End of file on the terminal!
-
- 
-Here is how much of TeX's memory you used:
- 2 strings out of 492942
- 14 string characters out of 3132734
- 56585 words of memory out of 3000000
- 4005 multiletter control sequences out of 15000+200000
- 3640 words of font info for 14 fonts, out of 3000000 for 9000
- 1141 hyphenation exceptions out of 8191
- 0i,0n,0p,1b,6s stack positions out of 5000i,500n,10000p,200000b,50000s
-!  ==> Fatal error occurred, no output PDF file produced!
