+++ tools/benchs/bench_avg.sh
